Add files via upload
This CSV file is a representative sample of the type of data used for analysis in the AMT_Pull_scpt project. It contains structured data similar to what would be processed by our Python scripts, serving as a model for understanding the input and output of our data pipeline. The structure of this sample mirrors actual datasets with anonymized values, illustrating potential variations and patterns in Germanium crystal growth metrics. This example is provided to demonstrate the capabilities of our parsing and visualization scripts and is not indicative of real proprietary information